+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+### Added
+
+### Fixed
+- `lt_ex_show_chip_id_and_fwver`: reboot back to Application mode in the end.
+
+### Removed
+
 ## [2.0.1]
 
 ### Added